requirements.txt
albumentations==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
dominate==2.6.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.19.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
networkx==2.8.2
numpy==1.22.3
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.1.1
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0
torchfile==0.1.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
visdom==0.1.8.9
wandb==0.12.16
websocket-client==1.3.2
